+++ src/screens/WalletScreen/Modals/SentModal/SentModal.types.ts
