+++ src/screens/WalletScreen/screens/Transactions/NoTransactionsYet/NoTransactionsYet.types.ts
